evdevvibrator: cleanup, also solves module loading problems on partial builds.
