--- modules/evdevvibrator.h
@@ -1,11 +1,0 @@
-#ifndef _EVDEV_VIBRA_H_
-#define _EVDEV_VIBRA_H_
-
-#include <glib.h>
-
-#define DEVICE "/dev/input/by-path/platform-vibrator-event"
-
-#define MCE_CONF_VIBRATOR_GROUP			"Vibrator"
-#define MCE_CONF_VIBRATOR_PATTERNS		"VibratorPatterns"
-
-#endif				/* _EVDEV_VIBRA_H_ */
